move MediaPlayerResponse under ECP
