+++ client/src/index.ts
@@ -23,5 +23,4 @@ export { Suitest, suitest };
 
 export * from './types/ActiveAppResponse';
 export * from './types/ConfigOptions';
-export * from './types/MediaPlayerResponse';
 export * from './types/OnDeviceComponentRequest';
